% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDIFtest_methods.R
\name{print.scDIFtest}
\alias{print.scDIFtest}
\title{Methods for class "scDIFtest"}
\usage{
\method{print}{scDIFtest}(x, item_selection = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{scDIFtest}}

\item{item_selection}{either \code{NULL} or an integer vector selecting the
item numbers. When \code{items = NULL} (the default), the DIF test
is done for all items.}

\item{...}{other argements passed to the method.}
}
\description{
Methods for class "scDIFtest"
}
